# specs/001-growth-check-webapp/contracts/api.yaml
openapi: 3.0.0
info:
  title: GrowthCheck API
  version: 1.0.0
  description: API for GrowthCheck Webapp (Next.js Frontend)
paths:
...
  /children:
    get:
      summary: List children (for Staff/Head)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: locationId
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated children list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChildIndex'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
...
  /children/{id}/measurements:
    post:
      summary: Add new measurement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeasurementInput'
      responses:
        '201':
          description: Created
...
components:
  schemas:
...
    ChildIndex:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        ageMonths:
          type: integer
        status:
          type: string
    MeasurementInput:
      type: object
      properties:
        height:
          type: number
        weight:
          type: number
        date:
          type: string
          format: date-time
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer